fix *prompb.x, move recording rules funcs
